Added additional command line arguments

+++ server/main/src/com/smartnote/server/util/FileUtils.java
@@ -1,7 +1,10 @@
 package com.smartnote.server.util;
 
 import java.io.File;
+import java.io.IOException;
 import java.nio.file.Files;
+import java.nio.file.InvalidPathException;
+import java.util.Objects;
 
 /**
  * File utilities.
@@ -15,9 +18,14 @@ public class FileUtils {
      * Deletes a file or directory. If the file is a directory, all files
      * in the directory will be deleted recursively.
      * 
-     * @param f The file to delete.
+     * @param f The file to delete. If <code>null</code>, nothing happens.
+     * @throws SecurityException If a security manager exists and its
+     *      <code>checkDelete</code> method denies delete access to the file or
+     *      <code>checkRead</code> method denies read access to the file.
      */
-    public static void deleteFile(File f) {
+    public static void deleteFile(File f) throws SecurityException {
+        if (f == null) return;
+
         if (f.isDirectory()) {
             for (File f2 : f.listFiles())
                 deleteFile(f2);
@@ -31,8 +39,12 @@ public class FileUtils {
      * in the directory will be deleted recursively.
      * 
      * @param path The path to the file to delete.
+     * @throws SecurityException If a security manager exists and its
+     *     <code>checkDelete</code> method denies delete access to the file or
+     *     <code>checkRead</code> method denies read access to the file.
      */
-    public static void deleteFile(String path) {
+    public static void deleteFile(String path) throws SecurityException {
+        if (path == null) return;
         deleteFile(new File(path));
     }
     
@@ -40,11 +52,16 @@ public class FileUtils {
      * Reads a file.
      * 
      * @param f The file to read.
-     * @return The contents of the file.
+     * @return The contents of the file. Never <code>null</code>.
      * 
-     * @throws Exception If an error occurs.
+     * @throws InvalidPathException If the path is invalid.
+     * @throws IOException If an I/O error occurs.
+     * @throws OutOfMemoryError If there is not enough memory to read the file.
+     * @throws SecurityException If a security manager exists and its
+     *       <code>checkRead</code> method denies read access to the file.
      */
-    public static String readFile(File f) throws Exception {
+    public static String readFile(File f) throws InvalidPathException,
+        IOException, OutOfMemoryError, SecurityException {
         return new String(Files.readAllBytes(f.toPath()));
     }
 
@@ -52,11 +69,16 @@ public class FileUtils {
      * Reads a file.
      * 
      * @param path The path to the file to read.
-     * @return The contents of the file.
+     * @return The contents of the file. Never <code>null</code>.
      * 
-     * @throws Exception If an error occurs.
+     * @throws InvalidPathException If the path is invalid.
+     * @throws IOException If an I/O error occurs.
+     * @throws OutOfMemoryError If there is not enough memory to read the file.
+     * @throws SecurityException If a security manager exists and its
+     *       <code>checkRead</code> method denies read access to the file.
      */
-    public static String readFile(String path) throws Exception {
+    public static String readFile(String path) throws InvalidPathException,
+        IOException, OutOfMemoryError, SecurityException {
         return readFile(new File(path));
     }
 
@@ -74,7 +96,16 @@ public class FileUtils {
         return filePath.startsWith(directoryPath);
     }
 
-    public static File getCanonicalFile(File f) {
+    /**
+     * Gets the canonical file of a file. If the canonical file cannot be
+     * retrieved, the absolute file is returned.
+     * 
+     * @param f The file.
+     * @return The canonical file.
+     * @throws SecurityException If a security manager exists and a system
+     * property cannot be accessed.
+     */
+    public static File getCanonicalFile(File f) throws SecurityException {
         try {
             return f.getCanonicalFile();
         } catch (Exception e) {
@@ -82,7 +113,16 @@ public class FileUtils {
         }
     }
 
-    public static String getCanonicalPath(File f) {
+    /**
+     * Gets the canonical path of a file. If the canonical path cannot be
+     * retrieved, the absolute path is returned.
+     * 
+     * @param f The file.
+     * @return The canonical path.
+     * @throws SecurityException If a security manager exists and a system
+     * property cannot be accessed.
+     */
+    public static String getCanonicalPath(File f) throws SecurityException {
         try {
             return f.getCanonicalPath();
         } catch (Exception e) {
@@ -90,6 +130,12 @@ public class FileUtils {
         }
     }
 
+    /**
+     * Creates a file path from a list of files.
+     * 
+     * @param files The files.
+     * @return The file path.
+     */
     public static File path(File... files) {
         File f = files[0];
         for (int i = 1; i < files.length; i++)
@@ -97,6 +143,12 @@ public class FileUtils {
         return f;
     }
 
+    /**
+     * Creates a file path from a list of files.
+     * 
+     * @param files The files.
+     * @return The file path.
+     */
     public static String path(String... files) {
         String f = files[0];
         for (int i = 1; i < files.length; i++)
@@ -110,8 +162,13 @@ public class FileUtils {
      * 
      * @param directory The directory.
      * @return The size of the directory.
+     * @throws NullPointerException If the directory is <code>null</code>.
+     * @throws SecurityException If a security manager exists and its
+     *    <code>checkRead</code> method denies read access to the file.
      */
-    public static long getDirectorySize(File directory) {
+    public static long getDirectorySize(File directory) throws NullPointerException, SecurityException {
+        Objects.requireNonNull(directory);
+
         long size = 0;
         for (File f : directory.listFiles()) {
             if (f.isDirectory())
